+++ Sources/StreamChat/Database/DTOs/MemberModelDTO.swift
@@ -23,8 +23,6 @@ class MemberDTO: NSManagedObject {
     @NSManaged var inviteAcceptedAt: DBDate?
     @NSManaged var inviteRejectedAt: DBDate?
     @NSManaged var isInvited: Bool
-    
-    @NSManaged var localInsertionAt: DBDate
 
     // MARK: - Relationships
 
@@ -53,8 +51,10 @@ extension MemberDTO {
         let request = NSFetchRequest<MemberDTO>(entityName: MemberDTO.entityName)
         request.predicate = NSPredicate(format: "ANY queries.queryHash == %@", query.queryHash)
         var sortDescriptors = query.sortDescriptors
-        // For consistent order, we need to have a sort descriptor which breaks ties
-        sortDescriptors.append(NSSortDescriptor(keyPath: \MemberDTO.localInsertionAt, ascending: true))
+        // For consistent order we need to have a sort descriptor which breaks ties
+        if !sortDescriptors.isEmpty, !sortDescriptors.contains(where: { $0.key == ChannelMemberListSortingKey.userId.rawValue }) {
+            sortDescriptors.append(ChannelMemberListSortingKey.userId.sortDescriptor(isAscending: true))
+        }
         request.sortDescriptors = sortDescriptors
         return request
     }
@@ -95,7 +95,6 @@ extension MemberDTO {
         let request = fetchRequest(id: memberId)
         let new = NSEntityDescription.insertNewObject(into: context, for: request)
         new.id = memberId
-        new.localInsertionAt = DBDate()
         return new
     }
 }